requirements.txt
streamlit==1.24.0
plotly==5.15.0
pandas==1.5.3
google-cloud-storage==2.9.0
mlflow==2.4.1
xgboost==1.7.6
python-dotenv==1.0.0
